requirements.txt
pip==24.0
setuptools==69.2.0
wheel==0.43.0
requests>=2.31.0

# framework: FastAPI
fastapi>=0.110.0
fastapi-sessions>= 0.3.2
asgi-correlation-id>=4.3.1
starlette~=0.36.3

# MongoDB handlers
pymongo >= 4.6.2
dnspython>=2.6.1
mongoengine>=0.28.2

chardet>=5.2.0

# PyDantic data validators
pydantic[email]>=2.6.4
pydantic-extra-types >= 2.6.0
pydantic-settings >= 2.2.1
pydantic_core>=2.16.3
phonenumbers >= 8.13.32

# Keys and Cryptography
pyjwt>=2.8.0
pycryptodome==3.20.0
cryptography>=38.0.0
jwcrypto >= 1.5.6

# Data for testing
faker>=24.3.0

# Keyborads
keyboard>=0.13.5

# Handling YAML like a BOSS... or for this logs!
PyYAML>=6.0.1

# UnitTesting
pytest>=8.1.1

# QR Tools
segno>=1.6.1

# Pillow for image rendering
pillow>=10.3.0

# AMQP for RABBITMQ
pika>=1.3.2

# Scrapper tools
beautifulsoup4 >= 4.12.3

# OdooRPC
odoorpc~=0.10.1